refactor: move column classes to column package & remove unused import and list
