+++ EmployeeManagementSystem/src/main/java/com/cra08/excellentcode/column/ColumnCl.java
@@ -1,4 +1,6 @@
-package com.cra08.excellentcode;
+package com.cra08.excellentcode.column;
+
+import com.cra08.excellentcode.Employee;
 
 public class ColumnCl implements IColumn {
     @Override
