+++ EmployeeManagementSystem/src/main/java/com/cra08/excellentcode/column/IColumn.java
@@ -1,4 +1,6 @@
-package com.cra08.excellentcode;
+package com.cra08.excellentcode.column;
+
+import com.cra08.excellentcode.Employee;
 
 public interface IColumn {
     Employee setValue(Employee employee, String value);
